cmd ref: move cobra to cmd as an app package
