+++ cmd/server/app/run.go
@@ -1,4 +1,4 @@
-package cmd
+package app
 
 import (
 	"github.com/spf13/cobra"
